Renamed topics Slides component into SlidesList and moved Slide component from app/components into SlidesList folder.
