+++ app/modules/topics/components/SlidesList/Slide/index.js
